Add package-lock.json file for project.
A package-lock.json file was added to define the exact versions of each dependency for the 'fls-start' project. This includes plugins for webpack, gulp and other dependencies. This will help to ensure that everyone working on this project is using the same versions of dependencies and reduce bugs that may occur due to differences in versions. The package-lock.json file is automatically generated and should be committed into the source control